move getVersionInfo
Expose it on index.js instead of juttle-service.js.
