+++ lib/juttle-service.js
@@ -4,7 +4,6 @@ var express = require('express');
 var logger = require('log4js').getLogger('juttle-service');
 var read_config = require('juttle/lib/config/read-config');
 var JuttleAdapters = require('juttle/lib/runtime/adapters');
-var getVersionInfo = require('./version');
 
 // Exposed handler for registering routes on an express app
 var addRoutes = require('./routes');
@@ -51,6 +50,5 @@ function run(options) {
 module.exports = {
     configure: configure,
     run: run,
-    getVersionInfo: getVersionInfo,
     addRoutes: addRoutes
 };
