+++ test/juttle-service.spec.js
@@ -10,8 +10,6 @@ var Promise = require('bluebird');
 var findFreePort = Promise.promisify(require('find-free-port'));
 var fs = Promise.promisifyAll(require('fs'));
 var fs_extra = Promise.promisifyAll(require('fs-extra'));
-var JuttleAdapters = require('juttle/lib/runtime/adapters');
-var sinon = require('sinon');
 
 var JSDP = require('juttle-jsdp');
 var moment = require('moment');
@@ -1875,50 +1873,15 @@ describe('Juttle Service Tests', function() {
         });
     });
 
-    describe('version', function() {
-        let adapterListStub;
-        before(function() {
-            adapterListStub = sinon.stub(JuttleAdapters, 'list');
-            adapterListStub.returns([
-                {
-                    adapter: 'elastic',
-                    version: '0.4.0'
-                },
-                {
-                    adapter: 'file',
-                    version: '0.5.1'
-                }
-            ]);
-        });
-
-        after(function() {
-            adapterListStub.restore();
-        });
-
-        function verifyVersionInfo(versionInfo) {
-            let VERSION_REGEX = /[0-9]+\.[0-9]+\.[0-9]+/;
-            expect(versionInfo['juttle-elastic-adapter']).to.equal('0.4.0');
-            expect(versionInfo['juttle']).to.match(VERSION_REGEX);
-            expect(versionInfo['juttle-service']).to.match(VERSION_REGEX);
-            expect(versionInfo['juttle-jsdp']).to.match(VERSION_REGEX);
-            // built in adapters (file in this case) should not be listed
-            expect(versionInfo).to.have.keys([
-                'juttle-elastic-adapter',
-                'juttle',
-                'juttle-service',
-                'juttle-jsdp'
-            ]);
-        }
-
-        it('juttle-service.getVersionInfo', function() {
-            let versionInfo = service.getVersionInfo();
-            verifyVersionInfo(versionInfo);
-        });
-
-        it('/version endpoint', function() {
+    describe('/version', function() {
+        it('contains component versions', function() {
             return chakram.get(juttleHostPort + '/version')
                 .then(function(response) {
-                    verifyVersionInfo(response.body);
+                    let versionInfo = response.body;
+                    let VERSION_REGEX = /[0-9]+\.[0-9]+\.[0-9]+/;
+                    expect(versionInfo['juttle']).to.match(VERSION_REGEX);
+                    expect(versionInfo['juttle-service']).to.match(VERSION_REGEX);
+                    expect(versionInfo['juttle-jsdp']).to.match(VERSION_REGEX);
                 });
         });
     });
